Seperate Calendar Repository to CalendarDateCalculator and self
