+++ app/src/main/java/com/example/todolist/repository/CalendarRepository.kt
@@ -2,6 +2,7 @@ package com.example.todolist.repository
 
 import android.util.Log
 import com.example.todolist.model.DayInfo
+import com.example.todolist.util.calculator.CalendarDateCalculator
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
 import com.google.firebase.database.database
@@ -11,29 +12,12 @@ class CalendarRepository {
     private val auth = Firebase.auth
     private val database = Firebase.database
     private val baseRef = database.getReference("Users/${auth.currentUser?.uid ?: ""}")
+    private val dateCalculator = CalendarDateCalculator()
 
     fun loadMonthData(yearMonth: YearMonth, callback: (List<DayInfo>) -> Unit) {
-        val firstDay = yearMonth.atDay(1)
-        val lastDay = yearMonth.atEndOfMonth()
-        val firstDayOfWeek = firstDay.dayOfWeek.value
-
-        val startDate = firstDay.minusDays((firstDayOfWeek - 1).toLong())
-        val endDate = lastDay.plusDays((7 - lastDay.dayOfWeek.value).toLong())
-
         baseRef.get()
         .addOnSuccessListener {
-            val dayInfoList = mutableListOf<DayInfo>()
-            var currentDate = startDate
-
-            while (!currentDate.isAfter(endDate)) {
-                dayInfoList.add(
-                    DayInfo(
-                        date = currentDate,
-                        isCurrentMonth = currentDate.month == yearMonth.month
-                    )
-                )
-                currentDate = currentDate.plusDays(1)
-            }
+            val dayInfoList = dateCalculator.calculateMonthDates(yearMonth)
             callback(dayInfoList)
         }
         .addOnFailureListener { exception ->
